% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbbottsCorrection.R
\name{AbbottsCorrection}
\alias{AbbottsCorrection}
\title{AbbottsCorrection}
\usage{
AbbottsCorrection(control.mortality, experimental.mortality,
  ci.quantile = 0.95)
}
\arguments{
\item{control.mortality}{Vector of fractional mortality in control group}

\item{experimental.mortality}{Vector of fractional mortality in treated group}

\item{ci.quantile}{\code{numeric} specifying the quantile for the confidence
intervals (default 0.95)}
}
\value{
Returns a named list of corrected mortality (\code{p.corr}) and
  confidence interval (\code{ci}) at specified \code{ci.quantile}
}
\description{
Perform Abbott's correction \strong{ref???}. \strong{add details}.
}

