% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupedAbbottsCorrectionDt.R
\name{GroupedAbbottsCorrectionDt}
\alias{GroupedAbbottsCorrectionDt}
\title{GroupedAbbottsCorrectionDt}
\usage{
GroupedAbbottsCorrectionDt(mortality.data, sample.group, ci.quantile = 0.95)
}
\arguments{
\item{mortality.data}{A \code{data.table} with colums \code{sample},
\code{dose} and \code{p} (uncorrected fractional mortality).}

\item{sample.group}{The group for this correction}

\item{ci.quantile}{\code{numeric} specifying the quantile for the confidence
intervals (default 0.95)}
}
\value{
\code{data.table} with colums \code{sample}, \code{dose},
  \code{p.corr} (corrected mortality) and \code{ci} (confidence interval at
  specified \code{ci.quantile})
}
\description{
\code{data.table} back-end to run groupued \code{AbbottsCorrection}s for
multiple samples. \strong{This corrects to \code{dose == 0}, should it be
possible to specify the base level? Add details}.
}

